move new edition card logic into edition provider
